requirements.txt
azure-functions
discord-interactions
